+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Fixed `IGMarkets::AccountMethods#activities` and `IGMarkets::AccountMethods#transactions` never returning
   if there were more than 500 activities or transactions in a single day during the specified period
 - Added `IGMarkets::Position#limited_risk_premium` and `IGMarkets::WorkingOrder#limited_risk_premium`
+- Supprt the `chart` transaction type
 
 ### 0.28 — February 24, 2017